+++ src/main/java/com/fesi/mukitlist/api/controller/web/kakao/client/KakaoTokenClient.java
@@ -1,6 +1,6 @@
-package com.fesi.mukitlist.api.controller.web.client;
+package com.fesi.mukitlist.api.controller.web.kakao.client;
 
-import com.fesi.mukitlist.api.controller.web.response.KakaoTokenResponse;
+import com.fesi.mukitlist.api.controller.web.kakao.response.KakaoTokenResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
